Create traefik-404.conf
This filter is realistically useful only within small Traefik deployments such as self-hosting. Use in large deployments will see legitimate users routinely blocked and systems intentionally open to the public should not even be considered.

It can however prove to be a useful tool in small user-base stable systems as it will routinely catch and block an impressive amount of unknown users and bots, many of which will be bad actors.